refactor: refactor contentParser to it's own file
